Widen mounts table statfs counts from uint32_t to uint64_t
Fixes #6549.

getMountedFilesystems() reads block and inode counts from statfs()
into MountInformation::StatFsInfo, but every field is declared
std::uint32_t and every assignment casts down with
static_cast<std::uint32_t>. On a 64-bit Linux host stats.f_blocks
(and f_bfree/f_bavail/f_files/f_ffree) come from a 64-bit
__fsblkcnt_t, so any filesystem with more than 2^32 blocks gets its
count silently wrapped before it ever reaches the mounts table.

The math lines up with the report: a 20T filesystem with 5362943488
4K blocks showed up as 1067976192, and 5362943488 mod 2^32 is
exactly 1067976192.

Widened all six StatFsInfo fields to std::uint64_t and the casts to
match. The mounts table already reports these columns as BIGINT and
Row values are stored as strings, so no other change is needed.

+++ osquery/filesystem/linux/mounts.h
@@ -18,22 +18,22 @@ namespace osquery {
 struct MountInformation final {
   struct StatFsInfo final {
     // Optimal transfer block size (statfs::f_bsize)
-    std::uint32_t block_size{0U};
+    std::uint64_t block_size{0U};
 
     // Total data blocks in file system (statfs::f_blocks)
-    std::uint32_t block_count{0U};
+    std::uint64_t block_count{0U};
 
     // Free blocks in filesystem (statfs::f_bfree)
-    std::uint32_t free_block_count{0U};
+    std::uint64_t free_block_count{0U};
 
     // Free blocks available to unprivileged user (statfs::f_bavail)
-    std::uint32_t unprivileged_free_block_count{0U};
+    std::uint64_t unprivileged_free_block_count{0U};
 
     // Total file nodes in filesystem (statfs::f_files)
-    std::uint32_t inode_count{0U};
+    std::uint64_t inode_count{0U};
 
     // Free file nodes in filesystem (statfs::f_ffree)
-    std::uint32_t free_inode_count{0U};
+    std::uint64_t free_inode_count{0U};
   };
 
   // Filesystem type
